server/requirements.txt
autopep8==2.1.0
channels==4.1.0
daphne==4.1.2
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
isort==5.13.2
pyopenssl==24.1.0
python-decouple==3.8
service-identity==24.1.0
stripe==9.7.0
zope.interface==6.4.post2
